[ir] Move HIR and MIR definitions to modules.
The `feldera-ir` crate now has a separate module for each type of IR.
